Update README.md
this is exlisevely for Macos setup, how it's gonna look depends on a lot of settings, I use wezterm , if somebody needs init.lua for linux don't hesitate to ask me i'll help as much as I can. PSit primarily for python programming, but you can use it for lua as well, don't forget to check mason, also check LspInfo to check if you have the right path to your environment.